Completed the three sub-tasks.
The getDataPoint() function now computes the price by taking the average of the ask and bid prices.

The getRatio() function now returns the ratio of the prices of the first stock to the second stock. It also includes an if condition to avoid the program from throwing a ZeroDivisionError.

The main part now stores and outputs the ratio properly.